Refactor importer tests to packages
